extract envelope to a new file
